+++ Cargo.toml
@@ -2,7 +2,7 @@
 resolver = "2"
 members = [
     "meilisearch-http",
-    "meilisearch-error",
+    "meilisearch-types",
     "meilisearch-lib",
     "meilisearch-auth",
     "permissive-json-pointer",
