+++ meilisearch-http/Cargo.toml
@@ -45,7 +45,7 @@ itertools = "0.10.3"
 jsonwebtoken = "8.0.1"
 log = "0.4.14"
 meilisearch-auth = { path = "../meilisearch-auth" }
-meilisearch-error = { path = "../meilisearch-error" }
+meilisearch-types = { path = "../meilisearch-types" }
 meilisearch-lib = { path = "../meilisearch-lib" }
 mime = "0.3.16"
 num_cpus = "1.13.1"
