+++ meilisearch-http/src/routes/indexes/documents.rs
@@ -6,10 +6,10 @@ use actix_web::{web, HttpRequest, HttpResponse};
 use bstr::ByteSlice;
 use futures::{Stream, StreamExt};
 use log::debug;
-use meilisearch_error::ResponseError;
 use meilisearch_lib::index_controller::{DocumentAdditionFormat, Update};
 use meilisearch_lib::milli::update::IndexDocumentsMethod;
 use meilisearch_lib::MeiliSearch;
+use meilisearch_types::error::ResponseError;
 use mime::Mime;
 use once_cell::sync::Lazy;
 use serde::Deserialize;
