+++ meilisearch-http/tests/index/create_index.rs
@@ -102,7 +102,7 @@ async fn error_create_with_invalid_index_uid() {
     let (response, code) = index.create(None).await;
 
     let expected_response = json!({
-        "message": "`test test#!` is not a valid index uid. Index uid can be an integer or a string containing only alphanumeric characters, hyphens (-) and underscores (_).",
+        "message": "invalid index uid `test test#!`, the uid must be an integer or a string containing only alphanumeric characters a-z A-Z 0-9, hyphens - and underscores _.",
         "code": "invalid_index_uid",
         "type": "invalid_request",
         "link": "https://docs.meilisearch.com/errors#invalid_index_uid"
