+++ meilisearch-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "meilisearch-error"
+name = "meilisearch-types"
 version = "0.28.0"
 authors = ["marin <postma.marin@protonmail.com>"]
 edition = "2021"
