+++ meilisearch-types/src/error.rs
